Add function to retrieve pub key bytes
It rests upon the user to store the pub key bytes from an attestation,
for use for verifying future assertions. Without any public means to
parse this from the attestation, it makes it the duty of the user to
write this logic for themselves, when the logic already exists within
the appattest-rs crate.

This exposes that logic as a public method and uses it in method that
logic came from to DRY things up a bit.