Add helper to create Assertion from CBOR string
This adds a helper method to generate an assertion from a raw CBOR
assertion string. This is useful because a caller may have requirements
that the assertion is encoded in a format other than plain base64, such
as base64 url encoded strings. This allows the caller to decode
appropriately on their end, and construct the Assertion object from the
raw CBOR string instead of enforcing all assertions must be plain base64
encoded.