Update assertion for separate hash and challenge
Previously the assertion would require client data bytes, hash it and
grab the challenge out of it. This is problematic for callers. The
client data object is often rich with other information, and would be
deserialized earlier in the process already. Re-serializing it could
result in a mismatch given changing json field order.

By taking the client hash and challenge separately, it makes it the
responsibility of the caller to hash the raw incoming josn payload and
then deserialize it, which is more appropriately done using something
like a custom actix extractor.

This updates the verify method, and the assertion example to take these
arguments independently, leaving it up to the caller to correctly hash
and deserialize the payload.