Don't check challenge on assertions
This removes checking challenges on assertions when we are verifying
against a list of app ids. This is to support our specific use case
where we want single routes that enforce app integrity checking, and we
already validated a client successfully, and stored their pub key. In
this case, the extra route to send a challenge doesn't provide extra
security, because the client is already signing the request hash itself
with a trusted pub key.